aoc_2025_08::models::nodes

Struct NodesList

pub struct NodesList {
    pub nodes: Vec<Node>,
    pub tree: KdTree<NodeDistanceType, usize, Node>,
}

A list of nodes that iterates over unique relations sorted by distance.

This struct owns the list of nodes as well as the KD-Tree used for efficient nearest-neighbour computation; it can produce a ClosestNeighboursIterator that yields unique relations between nodes in order of increasing distance.

Concept

Conceptually speaking, for any N number of nodes, there exists a complete matrix of N x N distances between each node and every other node. If one is to take this matrix,

remove all self-referential distances (i.e. distance from node A to node A),

remove all duplicate distances (i.e. A-B is the same as B-A), and then

sort the remaining distances in ascending order,

then this list is what this struct iterates over:

    p3 -> p6 = 2
    p4 -> p5 = 6
    p1 -> p2 = 7
    p1 -> p3 = 9
    p5 -> p6 = 9

However computing this complete matrix is O(N^2), which is infeasible for large N.

If we are to look at the problem differently, instead of the full matrix that we sort, we can have a sorted list of nearest-neighbour distances for each node:

    p1: p1 -> p2 = 7, p1 -> p3 = 9, ...
    p2: p2 -> p1 = 7, p2 -> p3 = 10, ...
    p3: p3 -> p6 = 2, p3 -> p1 = 9, ...
    p4: p4 -> p5 = 6, p4 -> p3 = 11, ...
    p5: p5 -> p4 = 6, p5 -> p6 = 9, ...
    p6: p6 -> p3 = 2, p6 -> p5 = 9, ...

Then we can scan the first nearest-neighbour of each node, the pop the smallest distance from that list, shifting the next nearest-neighbour of that node to the front:

    - popped p3 -> p6 = 2
    - p1: p1 -> p2 = 7, p1 -> p3 = 9, ...
    - p2: p2 -> p1 = 7, p2 -> p3 = 10, ...
    - p3: p3 -> p1 = 9, *p3 -> p2 = 10*, ...
    - p4: p4 -> p5 = 6, p4 -> p3 = 11, ...
    - p5: p5 -> p4 = 6, p5 -> p6 = 9, ...
    - p6: p6 -> p3 = 2, p6 -> p5 = 9, ...

There is no difference in the final sorted order of distances between this approach and the complete matrix approach. However, this approach allows for lazy evaluation of distances, and only requires the computation of nearest-neighbours for each node, when the node was popped: in the example above, we may not even know about p3 -> p2 = 10 at the time when we pop p3 -> p6 = 2, and we only compute it afterwards to fill the gap.

You may notice that the above examples have a lot of duplicate distances (e.g. p1 -> p2 = 7 and p2 -> p1 = 7). This can be avoided by only asking each node to find its nearest-neighbours where pN is higher than itself (i.e. only the bottom half of the distance matrix):

    p1: p1 -> p2 = 7, p1 -> p3 = 9, ...
    p2: p2 -> p3 = 10, ...
    p3: p3 -> p6 = 2, ...
    p4: p4 -> p5 = 6, ...
    p5: p5 -> p6 = 9, ...

p6 has no entries because it is the highest node. Any node that is only connected by nodes lower than itself will not have any entries as well.

This ensures our whole table only has <N-1 entries at any time, and they shall always be non-repeating. We can be assured that this produces the same result, because if one of the repeats was the smallest distance, its counterpart in the other direction would have been in the heap at the same time, so keeping both is redundant.

To summarize, this struct, when using Self::iter_closest_neighbours (which produces a ClosestNeighboursIterator), implements the following algorithm:

builds a KD-Tree from the input nodes for efficient nearest-neighbour computation,

for each node, finds its nearest-neighbour that has not already been paired with it (i.e. only the bottom half of the distance matrix), and

stores these relations in a min-heap sorted by distance,

when popping a relation from the heap, fans out from the node_a of that relation to find its next nearest-neighbour that has not already been paired with it, and pushes that new relation onto the heap, replacing the popped relation.

this continues until all unique relations have been popped from the heap, or some stopping condition is met, e.g. all nodes have been joined into a single graph.

Fields
nodes: Vec<Node>tree: KdTree<NodeDistanceType, usize, Node>
Implementations

impl NodesList

pub fn build_from(nodes: Vec<Node>) -> Result<Self>

Build a NodesList from a list of nodes.

This will use a KD-Tree to efficiently compute nearest neighbors. Then for each node, it will find its nearest neighbours that had not already been paired with it, and store the resulting relations in a min-heap sorted by distance.

This allows us to iterate over all unique nodes in order of increasing distance to nearest neighbour - which is useful in joining cluster of nodes into trees based on proximity.

pub fn build_from_text(input: &str) -> Result<Self>

Build a NodesList from a textual representation of nodes.

pub fn get_node_by_id(&self, node_id: usize) -> Option<&Node>

Get a reference to a node by its ID.

pub fn iter_closest_neighbours<'a>( &'a self, ) -> Result<ClosestNeighboursIterator<'a>>

Get an iterator over unique relations sorted by distance.

pub fn len(&self) -> usize

Get the number of nodes in this list.